# Conflicts:
#	app/Portal.Consultoras.ServiceHost/Web.config
#	app/Portal.Consultoras.Web/Controllers/AdministrarEstrategiaController.cs
#	app/Portal.Consultoras.Web/Controllers/BaseController.cs
#	app/Portal.Consultoras.Web/Portal.Consultoras.Web.csproj

+++ app/Portal.Consultoras.Web/Controllers/Estrategias/EstrategiaController.cs
@@ -186,7 +186,7 @@ namespace Portal.Consultoras.Web.Controllers.Estrategias
             }
             catch (Exception ex)
             {
-                LogManager.LogManager.LogErrorWebServicesBus(ex, userData.CodigoConsultora, userData.CodigoISO, ((int)tipoConsulta).ToString());
+                LogManager.LogManager.LogErrorWebServicesBus(ex, userData.CodigoConsultora, userData.CodigoISO, tipoConsulta.ToString());
                 return Json(new
                 {
                     success = false,
